Clean up animation and url structures for streamlined ux.
