+++ cave_core/views/site_views.py
@@ -1,21 +1,18 @@
 # Framework Imports
 from django.conf import settings
-from django.contrib.auth import login, authenticate, update_session_auth_hash, logout
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth.forms import PasswordChangeForm
 from django.shortcuts import render, redirect
-from django.contrib.auth.forms import AuthenticationForm
 from django.views.decorators.cache import cache_page
 from django.views.decorators.csrf import csrf_exempt
 
 
 # Internal Imports
 from cave_core import forms, models
-from cave_core.utils.wrapping import redirect_logged_in_user
 
+# Views
 
-# @cache_page(60 * 10)
-# @csrf_exempt
+@cache_page(60)
+@csrf_exempt
 def root_view(request):
     """
     Root view
@@ -32,13 +29,13 @@ def root_view(request):
         },
     )
 
-# Views
-@login_required(login_url="/auth/login/")
-def index(request):
-    """
-    Index View
 
-    Render the server configured index page
+@login_required(login_url="/auth/login/")
+def info(request):
+    """
+    Info View
+
+    Render the server configured info page
     """
     # print("\n\nIndex\n")
     globals = models.Globals.get_solo()
@@ -179,143 +176,3 @@ def profile(request):
                 "submit_button": "Update Profile",
             },
         )
-
-
-@login_required(login_url="/auth/login/")
-def change_password(request):
-    """
-    Change password view
-
-    Allows users to change their password
-    """
-    if request.method == "POST":
-        form = PasswordChangeForm(request.user, request.POST)
-        if form.is_valid():
-            user = form.save()
-            update_session_auth_hash(request, user)  # Important!
-            return redirect("/app/")
-    else:
-        form = PasswordChangeForm(request.user)
-    return render(
-        request,
-        "form.html",
-        {
-            "globals": models.Globals.get_solo(),
-            "access_dict": request.user.get_access_dict(),
-            "form": form,
-            "form_title": "Change Your Password",
-            "submit_button": "Change Password",
-        },
-    )
-
-@redirect_logged_in_user
-def signup(request):
-    """
-    User signup view
-
-    Users can create an account with this view
-    """
-    globals = models.Globals.get_solo()
-    if not globals.allow_anyone_create_user:
-        return redirect("/auth/login/")
-    if request.method == "POST":
-        form = forms.CreateUserForm(request.POST)
-        if form.is_valid():
-            new_user = form.save()
-            username = form.cleaned_data.get("username")
-            raw_password = form.cleaned_data.get("password1")
-            user = authenticate(username=username, password=raw_password)
-            login(request, user)
-            return redirect("/app/")
-    else:
-        form = forms.CreateUserForm()
-    return render(
-        request,
-        "form.html",
-        {
-            "globals": globals,
-            "form": form,
-            "form_title": "Create Account",
-            "submit_button": "Create",
-        },
-    )
-
-@redirect_logged_in_user
-def login_view(request):
-    """
-    User login view
-
-    Users can login to the site with this view
-    """
-    globals = models.Globals.get_solo()
-    if request.method == "POST":
-        form = AuthenticationForm(request, request.POST)
-        if form.is_valid():
-            username = form.cleaned_data.get("username")
-            password = form.cleaned_data.get("password")
-            user = authenticate(username=username, password=password)
-            if user is not None:
-                login(request, user)
-                next_url = request.POST.get("next_url")
-                if next_url == "None" or len(next_url) == 0:
-                    next_url = "/app/"
-                # Ensure next url ends with a trailing slash
-                if next_url[-1] != "/":
-                    next_url += "/"
-                return redirect(next_url)
-    else:
-        form = AuthenticationForm()
-    return render(
-        request,
-        "login.html",
-        {
-            "globals": globals,
-            "form": form,
-            "next_url": request.GET.get("next"),
-            "form_title": "Login",
-            "submit_button": "Login",
-        },
-    )
-
-def validate_email(request):
-    """
-    Site endpoint to validate an email code and redirect back to the index view
-
-    Requires:
-    - `code`:
-    ----- What: The randomly generated code used to validate the requesting user's email address
-    ----- Type: str
-
-    Example input (GET JSON):
-
-    -----------------------------------
-    {
-    "code":"myRandomCodeHere",
-    }
-    -----------------------------------
-    """
-    user = models.CustomUser.objects.filter(email_validation_code=request.GET.get("code")).first()
-    if user:
-        if user.email_validated == False:
-            user.email_validated = True
-            user.email_validation_code = None
-            user.save()
-        return redirect("/app/")
-    globals = models.Globals.get_solo()
-    return render(
-        request,
-        "validation_email_failed.html",
-        {
-            "globals": globals
-        },
-    )
-
-@login_required
-def user_logout(request):
-    """
-    Logout view
-
-    Allows users to logout of the site
-    """
-    logout(request)
-    return redirect("/auth/login/")
