+++ templates/app.html
@@ -8,7 +8,7 @@
 {% block content %}
   <div id="main-app-window-div" style="
     --widthMultiplier: 1;
-    height: calc( 100vh - 56px );
+    height: calc( 100vh - 58px );
     width: 100vw;
     overflow:hidden;
   ">
