+++ templates/base.html
@@ -14,7 +14,6 @@
   <!-- CSS -->
   <link rel="stylesheet" href="{% static 'css/materialicons.css' %}" type="text/css">
   <link rel="stylesheet" href="{% static 'css/bootstrap.min.css' %}" type="text/css" media="screen" />
-  <link rel="stylesheet" href="{% static 'css/animations.css' %}" type="text/css" media="screen" />
   <link rel="stylesheet" href="{% static 'css/style.css' %}" type="text/css" media="screen" />
   <!-- NOTE: This is loaded as a django template to pass db values directly into the script -->
   {% include "modules/base/style.html" %}
