+++ templates/modules/base/style.html
@@ -5,5 +5,8 @@
     background-image: linear-gradient( rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.6)), url({{ globals.site_background.url }});
     background-size: cover;
     background-position: center;
+    height: calc( 100vh - 58x );
+    margin: 0px;
+    padding: 0px;
   }
 </style>
